register_prefix refactoring
